Added project org.osate.aadl2.properties.  Moved AadlContribUtils from org.osate.aadl2.contrib to there.
