+++ analyses/org.osate.analysis.architecture/META-INF/MANIFEST.MF
@@ -19,9 +19,10 @@ Require-Bundle: org.osate.aadl2;bundle-version="[4.0.0,5.0.0)",
  org.eclipse.xtext;bundle-version="[2.20.0,3.0.0)",
  org.eclipse.emf.transaction;bundle-version="[1.4.0,2.0.0)",
  org.eclipse.ui.navigator;bundle-version="[3.9.0,4.0.0)",
- org.osate.aadl2.contrib;bundle-version="[1.2.0, 2.0.0)",
+ org.osate.aadl2.contrib;bundle-version="[1.2.0,2.0.0)",
  org.osate.contribution.sei;bundle-version="[1.2.0,2.0.0)",
- org.osate.pluginsupport;bundle-version="[7.1.0, 8.0.0)"
+ org.osate.pluginsupport;bundle-version="[7.1.0,8.0.0)",
+ org.osate.aadl2.properties;bundle-version="[1.0.0,2.0.0)"
 Bundle-ActivationPolicy: lazy
 Import-Package: org.osate.contribution.sei.names
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
