+++ analyses/org.osate.analysis.architecture/src/org/osate/analysis/architecture/PortConnectionConsistency.java
@@ -32,12 +32,12 @@ import org.osate.aadl2.contrib.aadlproject.SizeUnits;
 import org.osate.aadl2.contrib.communication.CommunicationProperties;
 import org.osate.aadl2.contrib.communication.RateSpec;
 import org.osate.aadl2.contrib.communication.RateSpec.RateUnit_FieldType;
-import org.osate.aadl2.contrib.util.AadlContribUtils;
 import org.osate.aadl2.instance.ConnectionInstance;
 import org.osate.aadl2.instance.ConnectionInstanceEnd;
 import org.osate.aadl2.instance.FeatureInstance;
 import org.osate.aadl2.instance.util.InstanceSwitch;
 import org.osate.aadl2.modelsupport.modeltraversal.AadlProcessingSwitchWithProgress;
+import org.osate.aadl2.properties.util.AadlContribUtils;
 import org.osate.contribution.sei.datamodel.DataModel;
 import org.osate.contribution.sei.sei.Sei;
 import org.osate.pluginsupport.properties.RealRange;
