+++ analyses/org.osate.analysis.flows/META-INF/MANIFEST.MF
@@ -20,7 +20,8 @@ Require-Bundle: org.osate.aadl2;bundle-version="[4.0.0,5.0.0)",
  org.osate.contribution.sei;bundle-version="[1.2.0,3.0.0)",
  org.osate.pluginsupport;bundle-version="[7.1.0,8.0.0)",
  org.osate.aadl2.contrib;bundle-version="[1.2.0,2.0.0)",
- org.osate.results;bundle-version="[3.0.0,4.0.0)"
+ org.osate.results;bundle-version="[3.0.0,4.0.0)",
+ org.osate.aadl2.properties;bundle-version="[1.0.0,2.0.0)"
 Bundle-ActivationPolicy: lazy
 Automatic-Module-Name: org.osate.analysis.flows
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
