+++ analyses/org.osate.analysis.resource.budgets/META-INF/MANIFEST.MF
@@ -21,7 +21,8 @@ Require-Bundle: org.eclipse.ui;bundle-version="[3.115.0,4.0.0)",
  org.osate.aadl2.contrib;bundle-version="[1.2.0,2.0.0)",
  org.osate.contribution.sei;bundle-version="[1.2.0,2.0.0)",
  org.osate.pluginsupport;bundle-version="[7.1.0,8.0.0)",
- org.osate.analysis.model;bundle-version="[1.0.0,2.0.0)"
+ org.osate.analysis.model;bundle-version="[1.0.0,2.0.0)",
+ org.osate.aadl2.properties;bundle-version="[1.0.0,2.0.0)"
 Bundle-ActivationPolicy: lazy
 Export-Package: org.osate.analysis.resource.budgets,
  org.osate.analysis.resource.budgets.busload,
