+++ analyses/org.osate.analysis.resource.budgets/src/org/osate/analysis/resource/budgets/logic/AbstractResourceAnalysis.java
@@ -31,12 +31,12 @@ import org.osate.aadl2.NamedElement;
 import org.osate.aadl2.contrib.aadlproject.SizeUnits;
 import org.osate.aadl2.contrib.aadlproject.TimeUnits;
 import org.osate.aadl2.contrib.timing.TimingProperties;
-import org.osate.aadl2.contrib.util.AadlContribUtils;
 import org.osate.aadl2.instance.ComponentInstance;
 import org.osate.aadl2.instance.FeatureCategory;
 import org.osate.aadl2.instance.FeatureInstance;
 import org.osate.aadl2.instance.SystemInstance;
 import org.osate.aadl2.instance.SystemOperationMode;
+import org.osate.aadl2.properties.util.AadlContribUtils;
 import org.osate.contribution.sei.sei.Instructionvolumeunits;
 import org.osate.contribution.sei.sei.ProcessorSpeedUnits;
 import org.osate.contribution.sei.sei.Sei;
