+++ analyses/org.osate.analysis.resource.budgets/src/org/osate/analysis/resource/budgets/logic/BusLoadAnalysis.java
@@ -34,7 +34,6 @@ import org.osate.aadl2.Element;
 import org.osate.aadl2.UnitLiteral;
 import org.osate.aadl2.contrib.aadlproject.DataRateUnits;
 import org.osate.aadl2.contrib.aadlproject.SizeUnits;
-import org.osate.aadl2.contrib.util.AadlContribUtils;
 import org.osate.aadl2.instance.ComponentInstance;
 import org.osate.aadl2.instance.ConnectionInstance;
 import org.osate.aadl2.instance.ConnectionInstanceEnd;
@@ -44,6 +43,7 @@ import org.osate.aadl2.instance.SystemInstance;
 import org.osate.aadl2.instance.SystemOperationMode;
 import org.osate.aadl2.modelsupport.modeltraversal.ForAllElement;
 import org.osate.aadl2.modelsupport.modeltraversal.SOMIterator;
+import org.osate.aadl2.properties.util.AadlContribUtils;
 import org.osate.aadl2.util.Aadl2Util;
 import org.osate.analysis.resource.budgets.busload.NewBusLoadAnalysis;
 import org.osate.ui.dialogs.Dialog;
