+++ analyses/org.osate.analysis.resource.management/META-INF/MANIFEST.MF
@@ -16,8 +16,9 @@ Require-Bundle: org.osate.analysis.binpacking;bundle-version="[3.0.0,4.0.0)",
  org.osate.aadl2.modelsupport;bundle-version="[6.0.0,7.0.0)",
  org.osate.ui;bundle-version="[6.2.0,7.0.0)",
  org.osate.xtext.aadl2.properties;bundle-version="[3.1.0,4.0.0)",
- org.osate.aadl2.contrib;bundle-version="[1.1.0, 2.0.0)",
- org.osate.pluginsupport;bundle-version="[7.0.0, 8.0.0)"
+ org.osate.aadl2.contrib;bundle-version="[1.1.0,2.0.0)",
+ org.osate.pluginsupport;bundle-version="[7.0.0,8.0.0)",
+ org.osate.aadl2.properties;bundle-version="[1.0.0,2.0.0)"
 Bundle-ActivationPolicy: lazy
 Automatic-Module-Name: org.osate.analysis.resource.management
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
