+++ analyses/org.osate.codegen.checker/META-INF/MANIFEST.MF
@@ -12,8 +12,9 @@ Require-Bundle: org.eclipse.ui;bundle-version="[3.115.0,4.0.0)",
  org.osate.aadl2.instantiation;bundle-version="[1.1.0,2.0.0)",
  org.osate.aadl2;bundle-version="[4.0.0,5.0.0)",
  org.eclipse.help;bundle-version="[3.6.0,4.0.0)",
- org.osate.aadl2.contrib;bundle-version="[1.1.0, 2.0.0)",
- org.osate.pluginsupport;bundle-version="[7.0.0, 8.0.0)"
+ org.osate.aadl2.contrib;bundle-version="[1.1.0,2.0.0)",
+ org.osate.pluginsupport;bundle-version="[7.0.0,8.0.0)",
+ org.osate.aadl2.properties;bundle-version="[1.0.0,2.0.0)"
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
 Import-Package: org.eclipse.core.internal.resources,
