+++ core/org.osate.aadl2.properties/.gitignore
@@ -1,0 +1,1 @@
+/bin/
