+++ core/org.osate.xtext.aadl2/META-INF/MANIFEST.MF
@@ -29,7 +29,8 @@ Require-Bundle: org.eclipse.xtext.generator;bundle-version="[2.20.0,3.0.0)";reso
  org.apache.commons.lang;bundle-version="[2.6.0,3.0.0)",
  org.osate.aadl2.contrib;bundle-version="[1.2.0,2.0.0)",
  org.osate.contribution.sei;bundle-version="[1.2.0,2.0.0)",
- org.osate.pluginsupport;bundle-version="[7.1.0,8.0.0)"
+ org.osate.pluginsupport;bundle-version="[7.1.0,8.0.0)",
+ org.osate.aadl2.properties;bundle-version="[1.0.0,2.0.0)"
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: org.osate.xtext.aadl2,
  org.osate.xtext.aadl2.formatting2.regionaccess,
